Refactor ReportingEngine API and Core
- Updated sourcelink JSON files to point to the latest commit.
- Modified project dependencies to include Microsoft.EntityFrameworkCore.Design and Microsoft.EntityFrameworkCore.SqlServer with version constraints.
- Removed Microsoft.AspNetCore.OpenApi dependency from project assets and cache.
- Changed OrganizationId type from int to long in Organization entity.
- Updated ITemplateService interface to accept long clientId instead of int.
- Refactored TemplateService to improve question retrieval logic and ensure proper mapping to DTOs.
- Updated assembly informational version attributes across multiple projects.
- Made various binary file updates in obj and bin directories for both ReportingEngine.API and ReportingEngine.Core.

+++ backend/ReportingEngine.API/Controllers/ReportsController.cs
@@ -22,8 +22,7 @@ namespace ReportingEngine.API.Controllers
         [HttpGet("clients")]
         public async Task<IActionResult> GetClients()
         {
-            // FIX: Changed OrganizationID to OrganizationId (standard scaffolding name)
-            // If this still fails, try just "Id"
+            // FIX: Changed OrganizationID -> OrganizationId
             var clients = await _context.Organizations
                 .Select(o => new { o.OrganizationId, o.Name })
                 .ToListAsync();
@@ -33,9 +32,8 @@ namespace ReportingEngine.API.Controllers
 
         // ENDPOINT 2: Get the Drag-and-Drop Questions (Schema)
         [HttpGet("schema/{clientId}")]
-        public async Task<IActionResult> GetSchema(int clientId)
+        public async Task<IActionResult> GetSchema(long clientId)
         {
-            // This will work now that we fixed the Interface in Step 1
             var data = await _templateService.GetQuestionsForClientAsync(clientId);
             return Ok(data);
         }
